add an history abstraction and local/remote implementations
working on loading history data to the report
